1-27-2018 DELETE AND INLINE EDIT +3
Implemented the DELETE feature. Turned "Actions" into a blue button for
each row that opens a context menu on left clicking the button -
currently there is one option for each row: "Delete." It asks "Are you
sure" after clicking "Delete" - if you say yes, it will delete the row
immediately, and it also deletes the correct tree from the database.